tourprocessor try catches implemented, error handling and messages displayed

+++ TourPlannerAPI/Controllers/TourController.cs
@@ -121,7 +121,8 @@ namespace TourPlannerAPI.Controllers
 
             var tourInfo = tour.TourInfo;
 
-            _context.TourInfo.Remove(tourInfo);
+            if(tourInfo != null)
+                _context.TourInfo.Remove(tourInfo);
 
             _context.Tours.Remove(tour);
             await _context.SaveChangesAsync();
